modify : edited folder structure of components
