+++ src/components/Register/FacebookButton.tsx
@@ -1,7 +1,7 @@
 // import { styled } from "@mui/material/styles";
 import { Button } from "@mui/material";
-import { Children, ReactNode } from "react";
-import GoogleIcon from "@mui/icons-material/Google";
+import { ReactNode } from "react";
+import FacebookRoundedIcon from "@mui/icons-material/FacebookRounded";
 import { styled } from "@mui/material/styles";
 
 const CustomButton = styled(Button)({
@@ -23,17 +23,22 @@ interface prop {
   children: ReactNode;
 }
 
-const GoogleButton = ({ children }: prop) => {
+const FacebookButton = ({ children }: prop) => {
   return (
     <CustomButton
       variant="contained"
-      startIcon={<GoogleIcon sx={{ color: "red" }} />}
+      startIcon={
+        <FacebookRoundedIcon
+          sx={{
+            color: "blue",
+          }}
+        />
+      }
       sx={{
         backgroundColor: "#E8F0F1",
         color: "black",
         width: { md: "59%", xs: "100%" },
         height: "45px",
-        fontSize: { md: "1.2rem", xs: "0.5rem" },
         fontFamily: "Arial",
         borderRadius: "20px",
         textDecoration: "none",
@@ -51,4 +56,4 @@ const GoogleButton = ({ children }: prop) => {
   );
 };
 
-export default GoogleButton;
+export default FacebookButton;
