+++ src/components/Register/GoogleButton.tsx
@@ -1,7 +1,7 @@
 // import { styled } from "@mui/material/styles";
 import { Button } from "@mui/material";
-import { Children, ReactNode } from "react";
-import FacebookRoundedIcon from "@mui/icons-material/FacebookRounded";
+import { ReactNode } from "react";
+import GoogleIcon from "@mui/icons-material/Google";
 import { styled } from "@mui/material/styles";
 
 const CustomButton = styled(Button)({
@@ -23,23 +23,16 @@ interface prop {
   children: ReactNode;
 }
 
-const FacebookButton = ({ children }: prop) => {
+const GoogleButton = ({ children }: prop) => {
   return (
     <CustomButton
       variant="contained"
-      startIcon={
-        <FacebookRoundedIcon
-          sx={{
-            color: "blue",
-          }}
-        />
-      }
+      startIcon={<GoogleIcon sx={{ color: "red" }} />}
       sx={{
         backgroundColor: "#E8F0F1",
         color: "black",
         width: { md: "59%", xs: "100%" },
         height: "45px",
-        fontSize: "1.2rem",
         fontFamily: "Arial",
         borderRadius: "20px",
         textDecoration: "none",
@@ -57,4 +50,4 @@ const FacebookButton = ({ children }: prop) => {
   );
 };
 
-export default FacebookButton;
+export default GoogleButton;
