+++ src/components/Register/MyInput.tsx
@@ -1,5 +1,4 @@
 import TextField from "@mui/material/TextField";
-import Box from "@mui/material/Box";
 interface prop {
   type: "string" | "number" | "password" | "email" | "date";
   name: string;
